refactor(requestId): Extract Constant Request ID
Remove duplicate reqId constants
